% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_devops_tokens.R
\name{browse_devops_tokens}
\alias{browse_devops_tokens}
\title{Open a browser where you create ICES Devop personal access tokens}
\usage{
browse_devops_tokens()
}
\value{
nothing - opens a web browser.
}
\description{
user must have approprate access priviladges
}
