% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devops_builds.R
\name{devops_builds}
\alias{devops_builds}
\title{Get a list of devops builds}
\usage{
devops_builds(..., all = FALSE)
}
\arguments{
\item{...}{additional arguments to GET}

\item{all}{show all columns, default = FALSE}
}
\value{
a data.frame
}
\description{
This will list all the builds you have read access too
}
