% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devops_clone.R
\name{devops_clone}
\alias{devops_clone}
\title{Perform a clone from the ICES devops git server}
\usage{
devops_clone(remote_url, local_dir = NULL, pat = devops_pat())
}
\arguments{
\item{remote_url}{the url of the devops git repo}

\item{local_dir}{the local directory you want to clone into}

\item{pat}{personal access token for devops}
}
\value{
a git2r repository object
}
\description{
user must have approprate access priviladges
}
