% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devops_get.R
\name{devops_get}
\alias{devops_get}
\title{Perform a GET request to devops server}
\usage{
devops_get(url, pat, ...)
}
\arguments{
\item{url}{the url of the devops api}

\item{pat}{devops personal access token}

\item{...}{additional arguments to GET}
}
\value{
http response object
}
\description{
Must be run in interactive mode.  User is promted for a password, and if successful
a Javascript Web Token (JWT) is saved to the project root (current working directory).
The token will last for 1 hour, after which the user needs to update the token using
this function again.
}
