% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devops_pat.R
\name{devops_pat}
\alias{devops_pat}
\title{Perform a push to the ICES devops git server}
\usage{
devops_pat()
}
\value{
the users devops paersonal access token
}
\description{
user must have approprate access priviladges
}
