% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devops_repos.R
\name{devops_repos}
\alias{devops_repos}
\title{Get a list of repositories}
\usage{
devops_repos(pat, ...)
}
\arguments{
\item{pat}{devops personal access token}

\item{...}{additional arguments to GET}
}
\value{
a data.frame
}
\description{
This will list all the repositories you have read access too
}
