% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icesDevops-package.R
\docType{package}
\name{icesDevops-package}
\alias{icesDevops-package}
\alias{icesDevops}
\title{Functions to link to the ICES Devops webservices.}
\description{
Functions to link to the ICES Devops webservices to allow
 users to interact with ICES TAF git repositoroes, and download
 summaries and data products from ICES TAF repositories.
}
\details{
\emph{Download data:}
\tabular{ll}{
  \code{\link{devops_url}}  \tab build url for ICES Devops api\cr
  \code{\link{devops_get}}  \tab perform GET request with PAT\cr
}
\emph{Authentication:}
\tabular{ll}{

}
}
\references{
ICES Devops: \url{https://devops.ices.dk}.
}
\author{
Colin Millar.
}
